Parametrized setup for measurements (node).

+++ trunk/firmware/MainBoard/CdmaTransceiver/SoftConsole/Teton_MSS_MSS_CM3_0/Teton_MSS_MSS_CM3_0_app/teton.c
@@ -1,6 +1,6 @@
 #include "teton.h"
 
-uint8_t packet_rate = 40; // FIXME
+float packet_rate = 40; // FIXME
 
 void Teton_init(void)
 {
